chore(perf tests): compare speeds of pyquery and beautiful soup
Listing the number of ratings and comments on a beer took around
0.8 seconds with BeautifulSoup. There are about 80900 beers on
the site... You do the math. The idea was to test pyquery to see
if perfs could be improved, even slightly.

Add a speedcompare script that is only a caller to other scripts
and compare their time.

Add logging modules with configuration in logging.json

Add a crawler based on pyquery. It still has some weird results
that I need to debug.

Finally, make scripts more robust (stop if data is not logical,
such as the number of beer links is different from that max we
extracted).